# contracts/openapi/public-v1.yaml
openapi: 3.1.0
info:
  title: Detection Scheduling Workbench API
  version: 1.0.0
  description: Human-approved scheduling API for an electrical-product testing center.
servers:
  - url: /api/v1
security:
  - OidcSessionCookie: []
paths:
...
  /schedule-previews:
    get: { operationId: listSchedulePreviews, responses: { '200': { $ref: '#/components/responses/SchedulePreviewPage' } } }
    post:
      operationId: createSchedulePreview
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { event_ids: { type: array, items: { type: string } } } } } } }
      responses: { '202': { $ref: '#/components/responses/SchedulePreviewResponse' } }
  /schedule-previews/{preview_id}:
    parameters: [ { $ref: '#/components/parameters/PreviewId' } ]
    get: { operationId: getSchedulePreview, responses: { '200': { $ref: '#/components/responses/SchedulePreviewResponse' } } }
...
  /schedule-previews/{preview_id}/approve:
    post:
      operationId: approveSchedulePreview
      parameters: [ { $ref: '#/components/parameters/PreviewId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/IfMatch' } ]
      responses: { '200': { $ref: '#/components/responses/SchedulePreviewResponse' }, '409': { $ref: '#/components/responses/Problem' }, '503': { $ref: '#/components/responses/Problem' } }
  /schedule-previews/{preview_id}/reject:
    post:
      operationId: rejectSchedulePreview
      parameters: [ { $ref: '#/components/parameters/PreviewId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/IfMatch' } ]
      responses: { '200': { $ref: '#/components/responses/SchedulePreviewResponse' }, '409': { $ref: '#/components/responses/Problem' } }
  /schedule-steps/{step_id}/start:
    patch:
      operationId: startScheduleStep
      parameters: [ { $ref: '#/components/parameters/StepId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/IfMatch' } ]
      responses: { '200': { $ref: '#/components/responses/ScheduleStepResponse' }, '409': { $ref: '#/components/responses/Problem' } }
  /schedule-steps/{step_id}/complete:
    patch:
      operationId: completeScheduleStep
      parameters: [ { $ref: '#/components/parameters/StepId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/IfMatch' } ]
      responses: { '200': { $ref: '#/components/responses/ScheduleStepResponse' }, '409': { $ref: '#/components/responses/Problem' } }
...
  /events/{event_id}:
    get: { operationId: getEvent, parameters: [ { $ref: '#/components/parameters/EventId' } ], responses: { '200': { $ref: '#/components/responses/EventResponse' } } }
...
  /events/{event_id}/close:
    post:
      operationId: closeEvent
      parameters: [ { $ref: '#/components/parameters/EventId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/IfMatch' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/EventCloseRequest' } } } }
      responses: { '200': { $ref: '#/components/responses/EventResponse' }, '409': { $ref: '#/components/responses/Problem' } }
  /events/{event_id}/acknowledge:
    post:
      operationId: acknowledgeEvent
      parameters: [ { $ref: '#/components/parameters/EventId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/IfMatch' } ]
      responses: { '200': { $ref: '#/components/responses/EventResponse' }, '409': { $ref: '#/components/responses/Problem' } }
...
  /notifications/{notification_id}/read:
    patch:
      operationId: readNotification
      parameters: [ { $ref: '#/components/parameters/NotificationId' }, { $ref: '#/components/parameters/IdempotencyKey' } ]
      responses: { '204': { description: Marked read } }
...
components:
  securitySchemes:
    OidcSessionCookie:
      type: apiKey
      in: cookie
      name: __Host-public_session
      description: OIDC-authenticated BFF session cookie for same-origin browser requests. The BFF sets this cookie after the OIDC login flow with HttpOnly, Secure, and SameSite protections; the React client does not attach a Bearer token and the browser sends the cookie automatically. The BFF validates the OIDC-authenticated session and derives actor identity, center scope, and roles exclusively from verified claims.
  parameters:
...
    PreviewId: { name: preview_id, in: path, required: true, schema: { type: string } }
    StepId: { name: step_id, in: path, required: true, schema: { type: string } }
    EventId: { name: event_id, in: path, required: true, schema: { type: string } }
    NotificationId: { name: notification_id, in: path, required: true, schema: { type: string } }
...
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1 } }
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string, minLength: 1 } }
  responses:
    Problem: { description: Problem details, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
    ScheduleStepResponse: { description: Schedule step, content: { application/json: { schema: { $ref: '#/components/schemas/ScheduleStep' } } } }
    SchedulePreviewResponse: { description: Preview, content: { application/json: { schema: { $ref: '#/components/schemas/SchedulePreview' } } } }
    SchedulePreviewPage: { description: Preview page, content: { application/json: { schema: { $ref: '#/components/schemas/SchedulePreviewPage' } } } }
    EventResponse: { description: Event, content: { application/json: { schema: { $ref: '#/components/schemas/Event' } } } }
...
  schemas:
    Problem: { type: object, required: [type, title, status], properties: { type: { type: string }, title: { type: string }, status: { type: integer }, detail: { type: string }, instance: { type: string } } }
...
    ScheduleStep: { type: object, required: [id, order_id, project_id, starts_at, ends_at, status, frozen], properties: { id: { type: string }, order_id: { type: string }, project_id: { type: string }, equipment_id: { type: string }, employee_ids: { type: array, items: { type: string } }, starts_at: { type: string, format: date-time }, ends_at: { type: string, format: date-time }, status: { type: string }, frozen: { type: boolean }, version: { type: integer } } }
    Schedule: { type: object, required: [version, steps], properties: { version: { type: integer }, steps: { type: array, items: { $ref: '#/components/schemas/ScheduleStep' } } } }
    SchedulePreview:
      type: object
      required: [id, status, algorithm_used, solver_status, fallback_used, fallback_reason, base_schedule_version, resource_snapshot_version, frozen_step_count, changed_step_count, delayed_order_count, weighted_delay_minutes, total_delay_minutes, blockers, changes, schedule, version]
      properties:
        id: { type: string }
        status: { type: string, enum: [pending_review, approved, rejected, conflicted, failed] }
        algorithm_used: { type: string, enum: [cp_sat, sla_fallback] }
        solver_status: { type: string }
        fallback_used: { type: boolean }
        fallback_reason: { type: [string, 'null'] }
        base_schedule_version: { type: integer }
        resource_snapshot_version: { type: integer }
        frozen_step_count: { type: integer }
        changed_step_count: { type: integer }
        delayed_order_count: { type: integer }
        weighted_delay_minutes: { type: integer }
        total_delay_minutes: { type: integer }
        blockers: { type: array, items: { type: object } }
        changes: { type: array, items: { type: object } }
        schedule: { $ref: '#/components/schemas/Schedule' }
        version: { type: integer }
      oneOf:
        - properties:
            fallback_used: { type: boolean, const: false }
            fallback_reason: { type: 'null' }
        - properties:
            fallback_used: { type: boolean, const: true }
            fallback_reason:
              type: string
              enum: [cp_sat_infeasible, cp_sat_timeout_without_feasible_solution, cp_sat_execution_error, input_size_protection_threshold_exceeded]
    SchedulePreviewPage: { type: object, required: [items, page, page_size, total], properties: { items: { type: array, items: { $ref: '#/components/schemas/SchedulePreview' } }, page: { type: integer }, page_size: { type: integer }, total: { type: integer } } }
    Event: { type: object, required: [id, event_type, entity_type, status, severity, version, occurred_at], properties: { id: { type: string }, event_type: { type: string }, entity_type: { type: string }, status: { type: string }, severity: { type: string }, entity_id: { type: string }, version: { type: integer }, occurred_at: { type: string, format: date-time }, payload: { type: object } } }
    EventCloseRequest: { type: object, required: [disposition], properties: { disposition: { type: string, minLength: 1, maxLength: 1000 } } }
...